be fix

+++ src/models/entry-model.mjs
@@ -1,5 +1,4 @@
-// Note: db functions are async and must be
-// called with await from the controller
+
 import promisePool from '../utils/database.mjs';
 
 const listAllEntries = async () => {
@@ -26,13 +25,12 @@ const listAllEntriesByUserId = async (id) => {
   }
 };
 
-const findEntryById = async (id) => {
+const findEntryById = async (id, userId) => {
   try {
     const [rows] = await promisePool.query(
-      'SELECT * FROM FeelingEntries WHERE entry_id = ?',
-      [id],
+      'SELECT * FROM FeelingEntries WHERE entry_id = ? AND user_id = ?',
+      [id, userId],
     );
-    // console.log('rows', rows);
     return rows[0];
   } catch (e) {
     console.error('error', e.message);
@@ -41,21 +39,13 @@ const findEntryById = async (id) => {
 };
 
 const addEntry = async (entry) => {
+  const {user_id, entry_date, mood, crying, breakdowns, notes} = entry;
   const sql = `INSERT INTO FeelingEntries
                (user_id, entry_date, mood, crying, breakdowns, notes)
                VALUES (?, ?, ?, ?, ?, ?)`;
-  const params = [
-    entry.user_id,
-    entry.entry_date,
-    entry.mood,
-    entry.crying,
-    entry.breakdowns,
-    entry.notes,
-  ];
+  const params = [user_id, entry_date, mood, crying, breakdowns, notes];
   try {
-    // change query method?
     const rows = await promisePool.query(sql, params);
-    // console.log('rows', rows);
     return {entry_id: rows[0].insertId};
   } catch (e) {
     console.error('error', e.message);
@@ -63,29 +53,20 @@ const addEntry = async (entry) => {
   }
 };
 
-const updateEntryById = async (entry) => {
+const updateEntryById = async (entryId, userId, entryData) => {
   try {
-    const sql =
-      `UPDATE FeelingEntries
-       SET entry_date=?, mood=?, crying=?, breakdowns=?, notes=?
-       WHERE entry_id=?`;
-    const params = [
-      entry.entry_date,
-      entry.mood,
-      entry.crying,
-      entry.breakdowns,
-      entry.notes,
-      entry.entry_id,
-    ];
+    const params = [entryData, entryId, userId];
+    const sql = promisePool.format(
+      `UPDATE FeelingEntries SET ? WHERE entry_id=? and user_id=?`, params,
+    );
     const [result] = await promisePool.query(sql, params);
-    // console.log(result);
+    console.log(result);
     if (result.affectedRows === 0) {
       return {error: 404, message: 'entry not found'};
     }
-    return {message: 'entry data updated', entry_id: entry.entry_id};
+    return {message: 'entry data updated', entry_id: entryId};
   } catch (error) {
     // fix error handling
-    // now duplicate entry error is generic 500 error, should be fixed to 400 ?
     console.error('updateEntryById', error);
     return {error: 500, message: 'db error'};
   }
@@ -93,10 +74,9 @@ const updateEntryById = async (entry) => {
 
 const deleteEntryById = async (id) => {
   try {
-    const sql = 'DELETE FROM FeelingEntries WHERE entry_id=?';
-    const params = [id];
+    const sql = 'DELETE FROM FeelingEntries WHERE entry_id=? AND user_id=?';
+    const params = [id, userId];
     const [result] = await promisePool.query(sql, params);
-    // console.log(result);
     if (result.affectedRows === 0) {
       return {error: 404, message: 'entry not found'};
     }
